requirements.txt
pandas
numpy
torch
transformers
scikit-learn
requests
PyYAML
ta
